Work on funny short bit problem
This was also causing issues in DSsim so I thought I should also remove funny short transect segments as well as the small polygons that overlapped with other valid coverered area polygons. Seems to only be an issue for zigzags but added the code to parallel designs too as a check. Reference #3