Reduced run time of donttest
In rhub testing it is run as donttest so all the examples must take <30 minutes to run or else the check fails. I have reduced the runtime of these examples but added comments to the user so that the know to increase the number of reps or reduce the spacing.